EpiLAMA-IL/scripts/inference/requirements.txt
accelerate==1.5.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.2.0
biopython==1.84
biotite==0.41.2
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
cloudpathlib==0.20.0
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
dill==0.4.0
einops==0.8.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.6.0
fsspec==2024.10.0
huggingface-hub==0.26.5
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lightning==2.5.1
lightning-utilities==0.14.3
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
mizani==0.11.4
mpmath==1.3.0
msgpack==1.1.0
msgpack-numpy==0.4.8
multidict==6.4.3
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
patchworklib==0.6.5
patsy==1.0.1
peft==0.15.1
peptides==0.3.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
plotnine==0.13.6
prompt_toolkit==3.0.48
propcache==0.3.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
six==1.17.0
stack-data==0.6.3
statsmodels==0.14.4
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
torchmetrics==1.7.1
torchtext==0.18.0
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
xformers==0.0.28.post3
yarl==1.20.0
esm==3.1.2
